% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnph.R
\name{coxnph_bart}
\alias{coxnph_bart}
\title{Additive Non-Proportional Hazards Model with Bayesian Additive Regression Trees (BART)}
\usage{
coxnph_bart(
  formula,
  data,
  pop_haz = NULL,
  test_data = NULL,
  num_tree = 50,
  k = 1,
  num_burn = 2500,
  num_thin = 1,
  num_save = 2500
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted. The response should be a `Surv` object
as created by the `survival::Surv` function.}

\item{data}{A data frame, list, or environment (or object coercible by `as.data.frame`
to a data frame) containing the variables in the model.}

\item{pop_haz}{(optional) A numeric vector specifying the population hazard. If `NULL`,
a relative survival model is not fit.}

\item{test_data}{(optional) A data frame for the test set. If `NULL`, the training data
is used for testing.}

\item{num_tree}{The number of trees to be used in the BART model. Defaults to 50.}

\item{k}{A parameter controlling the amount of regularization in the BART model.
Defaults to 1.}

\item{num_burn}{The number of MCMC iterations to discard as burn-in. Defaults to 2500.}

\item{num_thin}{The thinning parameter for the MCMC sampler. Defaults to 1.}

\item{num_save}{The number of MCMC samples to save. Defaults to 2500.}
}
\value{
Returns a fitted non-proportional hazards model object of class "coxnph_bart". The object is a list containing the following components.
\itemize{
}
}
\description{
This function fits a non-proportional hazards regression model with a BART
component. The hazard function is modeled as piecewise constant on time
intervals \eqn{0 = t_0 < t_1 < \cdots < t_B = \infty}, and is given by
}
\details{
\deqn{h(t \mid X) = h_P(t \mid X) + h_0(t) e^{g(X,t)}}

where:
\itemize{
 \item \eqn{h(t | X)}: is the hazard function at time \eqn{t} given covariate vector \eqn{X}.
 \item \eqn{h_P(t | x)}: is the population hazard rate, which is assumed known.
 \item \eqn{h_0(t)}: is the baseline hazard function at time \eqn{t} of the excess hazard, modeled as a piecewise constant function.
 \item \eqn{X}: is the covariate vector.
 \item \eqn{g(X,t)}: is the non-linear component estimated using BART, which is also assumed to be piecewise-constant as a function of $t$.
}
}
\examples{
\dontrun{
data("leuk_data")
}
}
