% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpe_bart.R
\name{coxpe_bart}
\alias{coxpe_bart}
\title{Additive Hazards Cox Proportional Hazards Model with Bayesian Additive Regression Trees (BART)}
\usage{
coxpe_bart(
  formula,
  data,
  pop_haz = NULL,
  test_data = NULL,
  num_tree = 50,
  k = 1,
  num_burn = 2500,
  num_thin = 1,
  num_save = 2500
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted. The response should be a `Surv` object
as created by the `survival::Surv` function.}

\item{data}{A data frame, list, or environment (or object coercible by `as.data.frame`
to a data frame) containing the variables in the model.}

\item{pop_haz}{(optional) A numeric vector specifying the population hazard. If `NULL`,
a relative survival model is not fit.}

\item{test_data}{(optional) A data frame for the test set. If `NULL`, the training data
is used for testing.}

\item{num_tree}{The number of trees to be used in the BART model. Defaults to 50.}

\item{k}{A parameter controlling the amount of regularization in the BART model.
Defaults to 1.}

\item{num_burn}{The number of MCMC iterations to discard as burn-in. Defaults to 2500.}

\item{num_thin}{The thinning parameter for the MCMC sampler. Defaults to 1.}

\item{num_save}{The number of MCMC samples to save. Defaults to 2500.}
}
\value{
Returns a fitted Cox proportional hazards model object with BART, of class `"coxpe_bart"`. The object is a list containing the following components:
\itemize{
\item \code{r_test}: A numeric matrix representing the samples of \eqn{g(X)} for each observation in the test set, which has been centered to have mean 0 over the training set.
\item \code{r_train}: A numeric matrix representing the samples of \eqn{g(X)} for each observation in the training set, which has been centered to have mean 0 over the training set.
\item \code{base_haz}: A numeric matrix representing the samples of the parameters for the piecewise constant hazard.
\item \code{lambda_test}: A numeric matrix representing the samples of \eqn{g(X)} for each observation in the test set, which has not been centered.
\item \code{lambda_train}: A numeric matrix representing the samples of \eqn{g(X)} for each observation in the training set, which has not been.
\item \code{base_haz}: A numeric matrix representing the samples of the parameters for the piecewise constant hazard, which accounts for centering of \eqn{g(X)}.
\item \code{hazard}: A numeric matrix representing the samples of the parameters for the piecewise constant hazard, which does not account for centering of \eqn{g(X)}.
\item \code{cum_hazard}: A numeric matrix representing the samples of the parameters for the piecewise constant cumulative hazard.
\item \code{counts}: A matrix containing the count of the number of times each predictor is used in the trees of the BART model, across all saved MCMC samples.
\item \code{sigma_lambda}: A numeric vector containing the posterior samples of the leaf parameter standard deviation for the BART component of the model.
\item \code{loglik}: A numeric vector representing the log-likelihood of the model at each saved MCMC iteration.
\item \code{time_grid}: grid representing the bins along which the hazard is piecewise constant.
\item \code{loglik_obs}: matrix containing the unit-level log-likelihood evaluations, which can be used with the loo function in the loo package to compute leave-one-out expected log predictive densities for model comparison.
}
}
\description{
This function fits a Cox proportional hazards regression models with a BART component.
The baseline hazard is modeled using a piecewise constant hazard function to obtain a
flexible model for the baseline hazard. The model is given by
}
\details{
\deqn{h(t | X) = h_P(t | X) + h_0(t) e^{g(X)}}

where:
\itemize{
 \item \eqn{h(t | X)}: is the hazard function at time \eqn{t} given covariate vector \eqn{X}.
 \item \eqn{h_P(t | x)}: is the population hazard rate, which is assumed known.
 \item \eqn{h_0(t)}: is the baseline hazard function at time \eqn{t} of the excess hazard, modeled as a piecewise constant function.
 \item \eqn{X}: is the covariate vector.
 \item \eqn{g(X)}: is the non-linear component estimated using BART.
}
}
\examples{
\dontrun{
data("leuk_data")
coxpe_bart(formula = Surv(event_time, status) ~ age + sex,
           data = leuk_data,
           pop_haz = leuk_data$haz_rate)
}
}
