% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leuk_data.R
\docType{data}
\name{leuk_data}
\alias{leuk_data}
\title{Leukemia Dataset}
\format{
A data frame with X rows and Y columns:
\describe{
  \item{var1}{Description of the first variable. A numeric vector.}
  \item{var2}{Description of the second variable. A factor with levels "level1" and "level2".}
  ... (repeat for other variables) ...
}
}
\source{
Description of where/how this dataset was obtained.
}
\usage{
data(leuk_data)
}
\description{
A dataset containing information on leukemia patients.
}
\examples{
data(leuk_data)
head(leuk_data)
}
\keyword{datasets}
