% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_coxph.R
\name{weibull_coxph}
\alias{weibull_coxph}
\title{Weibull Cox Proportional Hazards Regression Model}
\usage{
weibull_coxph(
  formula,
  data,
  pop_haz = NULL,
  test_data = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame, list, or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{pop_haz}{A numeric vector specifying the population hazard. If NULL,
a vector of zeros with length equal to the number of rows in the training
data is used, indicating that a standard survival model will be fit.}

\item{test_data}{A data frame for the test set. If NULL, the training data
is used for testing.}

\item{verbose}{A logical value. If TRUE, it prints out additional
information including a warning message for unimplemented functionality
related to test data evaluations.}

\item{...}{Additional arguments to be passed to the `rstan::sampling`
function.}
}
\value{
Returns a fitted Weibull regression model object from the `rstan`
package.
}
\description{
This function fits an additive/excess hazard regression model to survival
data using a Cox proportional hazards model with a Weibull baseline hazard.
The model can be mathematically formulated as:
}
\details{
\deqn{h(t | X) = h_P(t | X) + h_E(t | X)}

where:
\itemize{
\item \eqn{h(t | X)}: is the hazard function at time \eqn{t} given covariate vector \eqn{X}.
\item \eqn{h_P(t | X)}: is the population hazard, which is assumed known.
\item \eqn{h_E(t | X)}: is the excess hazard, which follows a Cox proportional hazards model with Weibull baseline hazard \eqn{h_0(t | X) \, e^{X^\top\beta}}.
\item \eqn{X}: is the covariate vector.
\item \eqn{\beta}: is the vector of coefficients.
}
}
\examples{
\dontrun{
data("leuk_data")
weibull_coxph(formula = Surv(event_time, status) ~ age + sex,
              pop_haz = leuk_data$haz_rate,
              data = leuk_data)
}
}
\seealso{
\code{\link[rstan]{sampling}}
}
